refactor: introduce cobra
