+++ go.mod
@@ -1,10 +1,8 @@
-module lamy
+module github.com/akihisa1210/lamy
 
 go 1.16
 
 require (
 	github.com/AlecAivazis/survey/v2 v2.2.9
-	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9 // indirect
-	golang.org/x/sys v0.0.0-20210119212857-b64e53b001e4 // indirect
-	golang.org/x/text v0.3.3 // indirect
+	github.com/spf13/cobra v1.2.1
 )
